requirements.txt
pytest-playwright
pydantic